fix(trusty-mpm): replace-by-identity for managed hooks to stop duplicate accumulation
Closes #2015
